refactor(web): split browser geolocation modules
